Support using server's alternate-access-address
Adds a new useAlternateAccessAddress config flag, which, when set to
true, instructs the client to use the alternate access address published
by the server(s).

This was _kind_of_ supported in client v3.4 and earlier but required
that the client's seed address list contains the complete list of the
alternate access address of all cluster nodes. The v3.4 and earlier
clients were not able to automatically detect new cluster hosts, not
included in the seed address list, using the alternate access address.

This partial support for alternate-access-address was broken in Node.js
client v3.5.0 (using C client v4.3.14) due to the introduction of a new
feature that automatically detects if the seed address used by the
client is a load balancer address. Due to this change, the v3.5.0 and
later client will always lookup the cluster nodes' actual access-address
and use that instead of the address used as the seed address.