Work-around for C client build problem
The C client SDK includes a modules/lua/src/luaconf.h.orig file, which
gets linked to modules/lua/src/luaconf.h as part of the build process.
However, the client's source archive includes a copy of the luaconf.h
file. (As a real file, not a symlink.) When the two files are extracted
from the source archive, it sometimes happens that the luaconfig.h.orig
file gets a filesystem timestamp that's later than the
filesystem timestamp of the luaconf.h file. (This is due to the
granularity of the filesytem timestamps, i.e. probably 1 second?) When
that happens, the C client build fails with the following error:

    (cd ../../modules/lua/src; ln -s luaconf.h.orig luaconf.h)
    ln: luaconf.h: File exists
    make[1]: Leaving directory '/src/aerospike-client-c/modules/mod-lua'

To avoid the problem (until the C client Makefile gets fixed in an
upcoming release), we can simply delete the luaconf.h file before
starting the build.