Add delete operation (#322)
* Add delete operation
* Test delete op both with and without read op

On server versions 4.7.0.7 as well as 4.8.0.2, performing a delete-all
operation on a data-in-memory namespace _may_ crash the server if the
deleted record has more than one bin. [AER-6174]
Due to slight differences in memory layout on the server side, the test
_without_ the additional read op, would cause the crash, while the same
test case _with_ the read op, would _not_ cause the crash.