Unref wakeup trigger while clients are not connected
After closing aerospike clients without releasing the event loop the
wakeup trigger in aerospike's event loop is keeping the node.js process
alive. This is problem for test suits running in mocha as mocha does not
call `process.exit()` by default, which has to be worked around.

This patch implements a cleaner behavior by keeping aerospike's event
loop alive while preventing it from keeping node's event loop alive.
Similar to the behavior when calling `process.exit()` the aerospike's
event loop is not released explicitly.

This also changes `Client::close()` default behavior to not release the
event loop, but to unref it. This is a better behavior for most use
cases.